# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CoronaMecca API",
        "description": "CoronaMecca API.",
        "contact": {
            "email": "dann4624@edu.sde.dk"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://vacapi.semeicardia.online/api"
        }
    ],
    "paths": {
...
        "/zones": {
            "get": {
                "tags": [
                    "Zones"
                ],
                "summary": "Get a list of Zones",
                "description": "Get a list of Zones",
                "operationId": "ZonesList",
                "responses": {
                    "200": {
                        "description": "List of Zones"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Zones"
                ],
                "summary": "Create an Zone",
                "description": "Create a Zone",
                "operationId": "ZonesCreate",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name of the Zone",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Zone Name"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Zone created"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/zones/deleted": {
            "get": {
                "tags": [
                    "Zones"
                ],
                "summary": "Get a list of deleted Zones",
                "description": "Get a list of deleted Zones",
                "operationId": "ZonesListDeleted",
                "responses": {
                    "200": {
                        "description": "List of deleted Zones"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/zones/{id}": {
            "get": {
                "tags": [
                    "Zones"
                ],
                "summary": "Get a specific Zone",
                "description": "Get a specific Zone",
                "operationId": "ZonesShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Zone",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Zone object"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Zones"
                ],
                "summary": "Update an Zone",
                "description": "Update an Zone",
                "operationId": "ZonesUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Zone",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "example": 1
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name of the Zone",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Zone Name"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Zone updated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Zones"
                ],
                "summary": "Delete a Zone",
                "description": "Delete a Zone",
                "operationId": "ZonesDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Zone",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Zone deleted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/zones/{id}/notify": {
            "post": {
                "tags": [
                    "Zones"
                ],
                "summary": "Email all 'Lager Chefer' with an error message",
                "description": "Email all 'Lager Chefer' with an error message",
                "operationId": "ZonesNotify",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Zone",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "example": 1
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Type af Notifikation",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Motion"
                        }
                    },
                    {
                        "name": "message",
                        "in": "query",
                        "description": "Optional Message",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Motion"
                        }
                    },
                    {
                        "name": "temperature",
                        "in": "query",
                        "description": "Temperature",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "example": "2.2"
                        }
                    },
                    {
                        "name": "humidity",
                        "in": "query",
                        "description": "Humidity",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "example": "50"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lager Chefer Notified"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/zones/{id}/force": {
            "delete": {
                "tags": [
                    "Zones"
                ],
                "summary": "Permanently Delete a Zone",
                "description": "Permanently Delete a Zone",
                "operationId": "ZonesDeleteForce",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Zone",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Zone permanently deleted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/zones/{id}/restore": {
            "put": {
                "tags": [
                    "Zones"
                ],
                "summary": "Restore a deleted Zone",
                "description": "Restore a deleted Zone",
                "operationId": "ZonesRestore",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Zone",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Zone restored"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/authenticate/zone": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Get API Key for Zone-controller by authenticating with a valid user",
                "description": "Get API Key for Zone-controller by authenticating with a valid user",
                "operationId": "AuthenticateZone",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "email",
                            "minimum": 5,
                            "example": "dann4624@edu.sde.dk"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "password",
                            "minimum": 5,
                            "example": "Pass.1234"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "API Token and User sent"
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}